+++ src/consumer/consumer.js
@@ -22,6 +22,7 @@ import loader from '../cli/loader';
 import { BEFORE_IMPORT_ACTION, BEFORE_MIGRATION } from '../cli/loader/loader-messages';
 import BitMap from './bit-map/bit-map';
 import ComponentMap from './bit-map/component-map';
+import { MissingBitMapComponent } from './bit-map/exceptions';
 import logger from '../logger/logger';
 import DirStructure from './dir-structure/dir-structure';
 import { getLatestVersionNumber, pathRelative } from '../utils';
@@ -101,9 +102,7 @@ export default class Consumer {
     } catch (err) {
       msg = msg
         ? format(msg, err)
-        : `Warning: Bit is not be able to run the bind command. Please install bit-${
-          err.lang
-        } driver and run the bind command.`;
+        : `Warning: Bit is not be able to run the bind command. Please install bit-${err.lang} driver and run the bind command.`;
       if (err instanceof DriverNotFound) {
         console.log(chalk.yellow(msg)); // eslint-disable-line
       }
@@ -473,7 +472,7 @@ export default class Consumer {
       }
       // don't write dists files for authored components as the author has its own mechanism to generate them
       // also, don't write dists file for imported component, unless the user used '--dist' flag
-      const writeDistsFiles = dist && origin === COMPONENT_ORIGINS.IMPORTED;
+      componentWithDeps.component._writeDistsFiles = dist && origin === COMPONENT_ORIGINS.IMPORTED;
       return componentWithDeps.component.write({
         bitDir,
         force,
@@ -484,7 +483,6 @@ export default class Consumer {
         consumerPath: this.getPath(),
         driver: this.driver,
         writeBitDependencies,
-        writeDistsFiles,
         dependencies: componentWithDeps.dependencies,
         componentMap
       });
@@ -560,9 +558,7 @@ export default class Consumer {
   moveExistingComponent(bitMap: BitMap, component: Component, oldPath: string, newPath: string) {
     if (fs.existsSync(newPath)) {
       throw new Error(
-        `could not move the component ${component.id} from ${oldPath} to ${
-          newPath
-        } as the destination path already exists`
+        `could not move the component ${component.id} from ${oldPath} to ${newPath} as the destination path already exists`
       );
     }
     const componentMap = bitMap.getComponent(component.id);
@@ -648,8 +644,20 @@ export default class Consumer {
     if (!componentFromModel) return includeNewComponents; // the component was never committed
     const latestVersionRef = componentFromModel.versions[componentFromModel.latest()];
     const versionFromModel = await this.scope.getObject(latestVersionRef.hash);
-    const componentFromFileSystem = await this.loadComponent(idParsed);
-    return this.isComponentModified(versionFromModel, componentFromFileSystem);
+    try {
+      const componentFromFileSystem = await this.loadComponent(idParsed);
+      return this.isComponentModified(versionFromModel, componentFromFileSystem);
+    } catch (err) {
+      if (
+        err instanceof MissingFilesFromComponent ||
+        err instanceof ComponentNotFoundInPath ||
+        err instanceof MissingBitMapComponent
+      ) {
+        // the file/s have been deleted or the component doesn't exist in bit.map file
+        return false;
+      }
+      throw err;
+    }
   }
 
   async commit(